refactor the unlocking of user and fix bug on workplaces due to endpoint bringing back wrong field
