+++ src/app/features/admin/search/search-for-group/search-for-group.component.html
@@ -46,6 +46,7 @@
                     [item]="item"
                     [workplaceIndex]="workplaceIndex"
                     [(navigateToWorkplace)]="navigateToWorkplace"
+                    [(unlockUser)]="unlockUser"
                   ></app-workplace-dropdown>
                 </td>
               </tr>
