+++ src/app/features/admin/search/search-for-workplace/search-for-workplace.component.html
@@ -75,6 +75,7 @@
                       [item]="item"
                       [workplaceIndex]="workplaceIndex"
                       [(navigateToWorkplace)]="navigateToWorkplace"
+                      [(unlockUser)]="unlockUser"
                     ></app-workplace-dropdown>
                   </td>
                 </tr>
