+++ src/app/features/admin/search/workplace-dropdown/workplace-dropdown.component.ts
@@ -2,7 +2,6 @@ import { Component, Input } from '@angular/core';
 import { WorkplaceSearchItem } from '@core/model/admin/search.model';
 import { DialogService } from '@core/services/dialog.service';
 import { SwitchWorkplaceService } from '@core/services/switch-workplace.service';
-import { AdminUnlockConfirmationDialogComponent } from '@shared/components/link-to-parent-cancel copy/admin-unlock-confirmation';
 
 @Component({
   selector: 'app-workplace-dropdown',
@@ -12,6 +11,7 @@ export class WorkplaceDropdownComponent {
   @Input() item: WorkplaceSearchItem;
   @Input() workplaceIndex: number;
   @Input() navigateToWorkplace: () => void;
+  @Input() unlockUser: (username: string, workplaceIndex: number, user: any) => void;
 
   constructor(private dialogService: DialogService, private switchWorkplaceService: SwitchWorkplaceService) {}
 
@@ -24,13 +24,6 @@ export class WorkplaceDropdownComponent {
 
   public unlockWorkplaceUser(username: string, workplaceIndex: number, userIndex: number, e: Event): void {
     e.preventDefault();
-    const data = {
-      username,
-      removeUnlock: () => {
-        this.item.users[userIndex].isLocked = false;
-      },
-    };
-
-    this.dialogService.open(AdminUnlockConfirmationDialogComponent, data);
+    this.unlockUser(username, workplaceIndex, this.item.users[userIndex]);
   }
 }
