+++ src/app/features/search/search.component.ts
@@ -8,7 +8,7 @@ import { DialogService } from '@core/services/dialog.service';
 import { SwitchWorkplaceService } from '@core/services/switch-workplace.service';
 import {
   AdminUnlockConfirmationDialogComponent,
-} from '@shared/components/link-to-parent-cancel copy/admin-unlock-confirmation';
+} from '@shared/components/admin-unlock-confirmation/admin-unlock-confirmation';
 
 @Component({
   selector: 'app-search',
