+++ src/app/features/search/search.module.ts
@@ -7,12 +7,16 @@ import { EmailTemplateResolver } from '@core/resolvers/admin/email-template.reso
 import { InactiveWorkplacesResolver } from '@core/resolvers/admin/inactive-workplaces.resolver';
 import { EmailCampaignService } from '@core/services/admin/email-campaign.service';
 import { DialogService } from '@core/services/dialog.service';
-import { AdminUnlockConfirmationDialogComponent } from '@shared/components/link-to-parent-cancel copy/admin-unlock-confirmation';
+import {
+  AdminUnlockConfirmationDialogComponent,
+} from '@shared/components/admin-unlock-confirmation/admin-unlock-confirmation';
 import { SharedModule } from '@shared/shared.module';
 
 import { CqcStatusChangeComponent } from './cqc-status-change/cqc-status-change.component';
 import { CqcStatusChangesComponent } from './cqc-status-changes/cqc-status-changes.component';
-import { SendEmailsConfirmationDialogComponent } from './emails/dialogs/send-emails-confirmation-dialog/send-emails-confirmation-dialog.component';
+import {
+  SendEmailsConfirmationDialogComponent,
+} from './emails/dialogs/send-emails-confirmation-dialog/send-emails-confirmation-dialog.component';
 import { EmailsComponent } from './emails/emails.component';
 import { ParentRequestComponent } from './parent-request/parent-request.component';
 import { ParentRequestsComponent } from './parent-requests/parent-requests.component';
